requirements.txt
sentence-transformers
fuzzywuzzy
google-generativeai
json
import-ipynb
python-dotenv
re
